Discover registry PEP 514 Pythons cross 32/64-bit
Fixes #11217

By default, a 64-bit uv does not see a 32-bit global (HKLM) installation of Python in the registry (https://github.com/astral-sh/uv/issues/11217). To work around this, we manually request both 32-bit and 64-bit access using registry access flags. The flags have no effect on 32-bit (https://stackoverflow.com/a/12796797/3549270).

This effect is that there is an asymmetry between discovery modes: For the registry-based discovery using PEP 514, we discover both 32-bit and 64-bit Pythons, while for managed installations, we are stricter and only discover those matching in bit-ness.